Many, many changes that are too long to write in a commit message I will outline everything in a gist
