+++ .gitignore
@@ -152,4 +152,4 @@ cython_debug/
 #.idea/
 
 # Miscellaneous
-test.py
+.py
