+++ utils/__init__.py
@@ -26,4 +26,14 @@ from .transformers import (
     NotBotRoleTransformer,
 )
 
+from errors import (
+    CaseAlreadyExists,
+    CaseDoesNotExist
+)
+from cases import (
+    Case,
+    Kick,
+    Ban,
+    CaseBase
+)
 # Files
