+++ utils/database/cases.py
@@ -1,3 +1,5 @@
+from typing import Any, Dict, List
+from uuid import UUID
 import discord
 
 # Discord Imports
@@ -30,7 +32,6 @@ class CasesDB(CoreDB):
         expires: datetime.datetime = None,
         **kwargs
     ):
-        expires = self.prepare_time(expires)
         async with self.pool.acquire() as conn:
             async with conn.transaction():
                 await conn.execute(
@@ -41,14 +42,14 @@ class CasesDB(CoreDB):
                     user_id,
                     moderator_id,
                     reason,
-                    created_at.replace(tzinfo=None),
-                    expires,
+                    self.prepare_time(created_at),
+                    self.prepare_time(expires),
                 )
                 if kwargs.get("return_case") == True:
                     cases = await conn.fetchval(
                         "SELECT COUNT(*) FROM cases WHERE guild_id=$1", guild_id
                     )
-                    log = await LogsDB(self.bot).get_moderation_logs(guild_id)
+                    log = await LogsDB(self.bot).get_config(guild_id)['moderation_logs']
                     return {"num": cases, "log_channel": log}
 
     async def get_case(self, guild_id: int, case_id: int):
@@ -71,11 +72,11 @@ class CasesDB(CoreDB):
                     return dict(case)
         return None
 
-    async def update_case(self, guild_id: int, case_id: int, reason: str):
+    async def update_case(self, case_id: int, reason: str):
         async with self.pool.acquire() as conn:
             async with conn.transaction():
-                case = await self.get_case(guild_id, case_id)
-                if isinstance(case, dict):
+                case = await self.check_case_exists(case_id)
+                if case:
                     await conn.execute(
                         "UPDATE cases SET reason=$1 WHERE case_id=$2", reason, case_id
                     )
@@ -87,11 +88,11 @@ class CasesDB(CoreDB):
                     return True
         return False
 
-    async def delete_case(self, guild_id: int, case_id: int):
+    async def delete_case(self, case_id: int):
         async with self.pool.acquire() as conn:
             async with conn.transaction():
-                case = await self.get_case(guild_id, case_id)
-                if isinstance(case, dict):
+                case = await self.check_case_exists(case_id)
+                if case:
                     await conn.execute("DELETE FROM cases WHERE case_id=$1", case_id)
                     return True
         return False
@@ -108,7 +109,7 @@ class CasesDB(CoreDB):
                     return [dict(case) for case in cases]
         return None
 
-    async def get_cases_by_member(self, guild_id: int, user_id: int):
+    async def get_cases_by_member(self, guild_id: int, user_id: int) -> List[Dict[str, Any]]:
         async with self.pool.acquire() as conn:
             async with conn.transaction():
                 cases = await conn.fetch(
@@ -151,3 +152,11 @@ class CasesDB(CoreDB):
                 if len(cases):
                     return len(cases)
         return None
+
+    async def check_case_exists(self, case_id:UUID):
+        async with self.pool.acquire() as conn:
+            async with conn.transaction():
+                case = await self.fetch_case(case_id)
+                if not case:
+                    return False
+                return case
