+++ utils/helpers.py
@@ -1,53 +1,14 @@
-"""
-The MIT License (MIT)
-
-Copyright (c) 2015 Rapptz
-
-Permission is hereby granted, free of charge, to any person obtaining a
-copy of this software and associated documentation files (the "Software"),
-to deal in the Software without restriction, including without limitation
-the rights to use, copy, modify, merge, publish, distribute, sublicense,
-and/or sell copies of the Software, and to permit persons to whom the
-Software is furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
-OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
-DEALINGS IN THE SOFTWARE.
-
-Snippets used from RoboDanny in 
-lines 103 - 185 https://github.com/Rapptz/RoboDanny/blob/rewrite/cogs/utils/time.py
-lines 200-264 https://github.com/Rapptz/RoboDanny/blob/rewrite/cogs/utils/formats.py
-lines 267-274 https://github.com/Rapptz/RoboDanny/blob/rewrite/cogs/admin.py
-"""
-
-
 import discord
-from discord.ext import commands
-from jishaku.functools import executor_function
 
 # Discord Imports
 
-from typing import Union, Iterable, Any, Dict, List, Optional
-import re
-from dateutil.relativedelta import relativedelta
-import parsedatetime as pdt
-import datetime
+from typing import Union, Iterable, Any, List
 import kimochi
 import humanize
-import pytz
 
 # Regular Imports
 
-from bot import ExultBot
 from .image_gen.emotion import Marriage
-from .database import ServerStatsDB, CustomCommandsDB
 
 # Local Imports
 
@@ -61,148 +22,6 @@ class HexType:
 
     def __str__(self):
         return hex(self.val)
-
-
-def embed_builder(
-    *,
-    title: str = None,
-    description: str = None,
-    colour: HexType = ExultBot.red,
-    timestamp: bool = None,
-    author: Union[list, str] = None,
-    footer: Union[list, str] = None,
-    thumbnail: str = None,
-    image: str = None,
-    fields: list = None,
-    url: str = None,
-):
-    embed = discord.Embed()
-    if title:
-        embed.title = title
-    if description:
-        embed.description = description
-    if timestamp:
-        embed.timestamp = timestamp
-    embed.colour = colour
-    if thumbnail:
-        embed.set_thumbnail(url=thumbnail)
-    if image:
-        embed.set_image(url=image)
-    if url:
-        embed.url = url
-    if author:
-        if isinstance(author, list):
-            embed.set_author(icon_url=author[0], name=author[1])
-        elif isinstance(author, str):
-            embed.set_author(name=author)
-    if footer:
-        if isinstance(footer, list):
-            embed.set_footer(icon_url=footer[0], text=footer[1])
-        elif isinstance(footer, str):
-            embed.set_footer(text=footer)
-    if fields:
-        for field in fields:
-            try:
-                embed.add_field(name=field[0], value=field[1], inline=field[2])
-            except IndexError:
-                embed.add_field(name=field[0], value=field[1])
-    return embed
-
-
-class ShortTime:
-    compiled = re.compile(
-        """(?:(?P<years>[0-9])(?:years?|y))?             # e.g. 2y
-                             (?:(?P<months>[0-9]{1,2})(?:months?|mo))?     # e.g. 2months
-                             (?:(?P<weeks>[0-9]{1,4})(?:weeks?|w))?        # e.g. 10w
-                             (?:(?P<days>[0-9]{1,5})(?:days?|d))?          # e.g. 14d
-                             (?:(?P<hours>[0-9]{1,5})(?:hours?|h))?        # e.g. 12h
-                             (?:(?P<minutes>[0-9]{1,5})(?:minutes?|m))?    # e.g. 10m
-                             (?:(?P<seconds>[0-9]{1,5})(?:seconds?|s))?    # e.g. 15s
-                          """,
-        re.VERBOSE,
-    )
-
-    def __init__(self, argument, *, now=None):
-        match = self.compiled.fullmatch(argument)
-        if match is None or not match.group(0):
-            raise commands.BadArgument("invalid time provided")
-
-        data = {k: int(v) for k, v in match.groupdict(default=0).items()}
-        now = now or datetime.datetime.now(datetime.timezone.utc)
-        self.dt = now + relativedelta(**data)
-
-    @classmethod
-    async def convert(cls, ctx, argument):
-        return cls(argument, now=ctx.message.created_at)
-
-
-class HumanTime:
-    calendar = pdt.Calendar(version=pdt.VERSION_CONTEXT_STYLE)
-
-    def __init__(self, argument, *, now=None):
-        now = now or datetime.datetime.utcnow()
-        dt, status = self.calendar.parseDT(argument, sourceTime=now)
-        if not status.hasDateOrTime:
-            raise commands.BadArgument(
-                "Make sure you format your duration correctly! `(e.g. 5 hours 30 minutes)`"
-            )
-
-        if not status.hasTime:
-            dt = dt.replace(
-                hour=now.hour,
-                minute=now.minute,
-                second=now.second,
-                microsecond=now.microsecond,
-            )
-
-        self.dt = dt
-        self._past = dt < now
-
-    @classmethod
-    async def convert(cls, ctx, argument):
-        if isinstance(ctx, discord.Interaction):
-            return cls(argument, now=ctx.created_at)
-        elif isinstance(ctx, commands.Context):
-            return cls(argument, now=ctx.message.created_at)
-
-
-class Time(HumanTime):
-    def __init__(self, argument, *, now=None):
-        try:
-            o = ShortTime(argument, now=now)
-        except:
-            super().__init__(argument)
-        else:
-            self.dt = o.dt
-            self._past = False
-
-
-class FutureTime(Time):
-    def __init__(self, argument, *, now=None):
-        argument_list = list(argument)
-        if " " not in argument_list:
-            pos = -1
-            for c in argument_list:
-                pos += 1
-                if str(c).isalpha():
-                    break
-            argument_list.insert(pos, " ")
-            argument = "".join([z for z in argument_list])
-        super().__init__(argument, now=now)
-
-        if self._past:
-            raise commands.BadArgument("This time is in the past.")
-
-
-def time_handler(time) -> datetime.datetime:
-    time = FutureTime(time).dt
-    try:
-        localised = ExultBot.time(time)
-    except:
-        raise commands.BadArgument(
-            "Make sure you format your duration correctly! `(e.g. 5 hours 30 minutes)`"
-        )
-    return localised
 
 
 class plural:
@@ -358,247 +177,3 @@ class Emotions:
         embed.set_image(url="attachment://marry.png")
         print("completed marriage image")
         return {"embed": embed, "file": file}
-
-
-class ServerUtils:
-    def __init__(
-        self,
-        category: discord.CategoryChannel,
-        timezone: str = None,
-        milestone: int = None,
-    ):
-        self.bot = None
-        self.db = None
-        self.category = category
-        self.func_dict = {
-            "Timezone": [self.timezone_channel, timezone],
-            "MemberCount": self.member_channel,
-            "Channels": self.channel_channel,
-            "Status Counter": self.member_count_channel,
-            "Milestone": [self.milestone_channel, milestone],
-        }
-        self.funcs = ["Timezone", "Member", "Channels", "Status Counter", "Milestone"]
-
-        self.to_be_uploaded = {
-            "Timezone": None,
-            "MemberCount": None,
-            "Channels": None,
-            "Status Counter": None,
-            "Milestone": None,
-        }
-
-    def overwrites(self, ctx: discord.Interaction):
-        overwrites = {
-            ctx.guild.default_role: discord.PermissionOverwrite(
-                connect=False, view_channel=True
-            )
-        }
-        return overwrites
-
-    @staticmethod
-    def guild_milestone(guild: discord.Guild, milestone: Optional[int] = None) -> int:
-        def gtx(num: int, *, bound: int) -> int:
-            if num % bound == 0:
-                return num + bound
-            return (num - (num % bound)) + bound
-
-        def bound(_num: int) -> int:
-            _num = milestone if milestone else _num
-            return int("1" + "0" * (len(str(_num)) - 1))
-
-        __temp = gtx(guild.member_count, bound=bound(guild.member_count))
-
-        if __temp < 100:
-            return 100
-        return gtx(guild.member_count, bound=bound(guild.member_count))
-
-    @staticmethod
-    def guild_member_count(guild: discord.Guild) -> int:
-        online = 0
-        idle = 0
-        dnd = 0
-        offline = 0
-        for m in guild.members:
-            if m.status == discord.Status.online:
-                online += 1
-            elif m.status == discord.Status.idle:
-                idle += 1
-            elif m.status == discord.Status.dnd:
-                dnd += 1
-            else:
-                offline += 1
-        return f"\U0001f7e2{online}\U0001f7e1{idle}\U0001f534{dnd}\U000026ab{offline}"
-
-    @staticmethod
-    def channel_count(guild: discord.Guild):
-        return f"\U0001f4dc {len(guild.text_channels)} \U0001f509 {len(guild.voice_channels)}"
-
-    @staticmethod
-    def get_time(country: str):
-        try:
-            timezone = pytz.timezone(country)
-            now = datetime.datetime.now(timezone)
-            return now.strftime("%H:%M")
-        except:
-            raise commands.BadArgument("Invalid timezone")
-
-    async def timezone_channel(
-        self,
-        ctx: discord.Interaction,
-        timezone: str = None,
-    ):
-        if timezone is None:
-            return None
-        channel = await self.category.create_voice_channel(
-            name=f"\U000023f0 {self.get_time(timezone)}",
-            position=0,
-            overwrites=self.overwrites(ctx),
-        )
-        self.to_be_uploaded["Timezone"] = [channel.id, timezone]
-
-    async def member_channel(self, ctx: discord.Interaction):
-        channel = await self.category.create_voice_channel(
-            name=f"\U0001fac2 Members {ctx.guild.member_count}",
-            position=1,
-            overwrites=self.overwrites(ctx),
-        )
-        self.to_be_uploaded["MemberCount"] = channel.id
-
-    async def channel_channel(self, ctx: discord.Interaction):
-        channel = await self.category.create_voice_channel(
-            name=self.channel_count(ctx.guild),
-            position=2,
-            overwrites=self.overwrites(ctx),
-        )
-        self.to_be_uploaded["Channels"] = channel.id
-
-    async def member_count_channel(self, ctx: discord.Interaction):
-        channel = await self.category.create_voice_channel(
-            name=self.guild_member_count(ctx.guild),
-            position=3,
-            overwrites=self.overwrites(ctx),
-        )
-        self.to_be_uploaded["Status Counter"] = channel.id
-
-    async def milestone_channel(self, ctx: discord.Interaction, milestone: int) -> None:
-        channel = await self.category.create_voice_channel(
-            name=f"\U0001f4c8 {ctx.guild.member_count}/{self.guild_milestone(ctx.guild, milestone)}",
-            position=4,
-            overwrites=self.overwrites(ctx),
-        )
-        self.to_be_uploaded["Milestone"] = [channel.id, milestone]
-
-    async def create_channels(self, ctx: discord.Interaction, types: list):
-        for _type in types:
-            await self.func_dict[_type][0](ctx, self.func_dict[_type][1]) if isinstance(
-                self.func_dict[_type], list
-            ) else await self.func_dict[_type](ctx)
-        await self.upload_to_db(ctx)
-
-    async def upload_to_db(self, ctx: discord.Interaction):
-        if not self.bot:
-            self.bot = ctx.client
-        if not self.db:
-            self.db = ServerStatsDB(self.bot)
-        time_stats = self.to_be_uploaded.get("Timezone")
-        member_stats = self.to_be_uploaded.get("MemberCount")
-        channel_stats = self.to_be_uploaded.get("Channels")
-        status_stats = self.to_be_uploaded.get("Status Counter")
-        milestone_stats = self.to_be_uploaded.get("Milestone")
-        if time_stats:
-            await self.db.upload_timezone(
-                ctx.guild.id, channel_id=time_stats[0], timezone=time_stats[1]
-            )
-        if member_stats:
-            await self.db.upload_membercount(ctx.guild.id, member_stats)
-        if channel_stats:
-            await self.db.upload_channels(ctx.guild.id, channel_stats)
-        if status_stats:
-            await self.db.upload_statuses(ctx.guild.id, status_stats)
-        if milestone_stats:
-            await self.db.upload_milestone(
-                ctx.guild.id,
-                channel_id=milestone_stats[0],
-                milestone=milestone_stats[1],
-            )
-
-
-class CommandUtils:
-    def __init__(self, bot: commands.Bot) -> None:
-        self.bot: commands.Bot = bot
-        self.db = CustomCommandsDB(bot)
-        self.cache: Dict[Union[str, int], Any] = {}
-
-    def checks(self, ctx: Union[discord.Interaction, commands.Context]) -> bool:
-        if not ctx.guild and not (
-            ctx.author.bot if isinstance(ctx, commands.Context) else ctx.user.bot
-        ):
-            return False
-        return True
-
-    async def get_prefix(
-        self, ctx: Union[commands.Context, discord.Message]
-    ) -> Optional[str]:
-        pre = await self.bot.get_prefix(
-            ctx.message if isinstance(ctx, commands.Context) else ctx
-        )
-        return pre[0] if isinstance(pre, list) else pre
-
-    async def create_command(
-        self,
-        ctx: Union[discord.Interaction, commands.Context],
-        name: str,
-        response: str,
-    ) -> None:
-        if self.checks(ctx):
-            if ctx.guild is not None:
-                resp = {
-                    name: {
-                        "guild_id": ctx.guild.id,
-                        "command_name": name,
-                        "command_text": response,
-                        "command_creator": ctx.author.id
-                        if isinstance(ctx, commands.Context)
-                        else ctx.user.id,
-                        "created_id": ctx.message.created_at.timestamp()
-                        if isinstance(ctx, commands.Context)
-                        else ctx.created_at.timestamp(),
-                    }
-                }
-                try:
-                    self.cache[ctx.guild.id].update(resp)
-                except KeyError:
-                    self.cache[ctx.guild.id] = {}
-                    self.cache[ctx.guild.id].update(resp)
-
-    def cache_command(self, command: dict):
-        guild_id = command.get("guild_id")
-        name = command.get("command_name")
-        response = command.get("command_text")
-        creator = command.get("command_creator")
-        created_at = command.get("created_at")
-        resp = {
-            name: {
-                "guild_id": guild_id,
-                "command_name": name,
-                "command_text": response,
-                "command_creator": creator,
-                "created_id": created_at,
-            }
-        }
-        try:
-            self.cache[guild_id].update(resp)
-        except KeyError:
-            self.cache[guild_id] = {}
-            self.cache[guild_id].update(resp)
-
-    async def get_command(
-        self, ctx: Union[discord.Interaction, commands.Context], name: str
-    ) -> Optional[Dict[str, Any]]:
-        if self.checks(ctx):
-            if ctx.guild is not None:
-                return self.cache[ctx.guild.id].get(name, "Not found..")
-
-    async def upload_db(self) -> None:
-        for guild in list(self.cache):
-            await self.db.upload_commands(self.cache[guild])
